feat: complete notification-system + risk-engine-v2 (16 issues)
- RE-001: PositionSizer + SlippageEstimator

- RE-002: TrailingStopManager (breakeven + trailing)

- RE-003: MarketRegimeDetector + TimeFilter

- RE-004: LMT Orders + API auto-conversion

- RE-005: Order Lifecycle State Machine (DB migration)

- RE-006: Real Fill Price Tracking (FillTracker)

- RE-007: ATR-Based Adaptive Stop Loss

- RE-008: Partial Exits (50% at 1.5x SL)

- RE-009: Re-entry Rules + Cooldown

- RE-010: ML Signal Filter (logistic regression pre-filter)

- RE-011: Drawdown Recovery Strategy

- NS-001: NotificationThrottler + NotificationQueue

- NS-002: ApprovalManager (async callbacks)

- NS-003: NotificationPolicy + DigestGenerator

- NS-004: OrderExecutionMonitor + fill verification

- NS-005: OrderDeduplication + PreflightChecker

- Push test coverage from 65% to 80% (697 tests, 0 failures)

- Add scikit-learn for SignalFilter model training

+++ app/risk/validator.py
@@ -16,7 +16,7 @@ ALLOWED_ORDER_TYPES = {"MKT", "LMT"}
 class ValidationResult:
     approved: bool
     reasons: list[str] = field(default_factory=list)
-    position_size_units: int = 0
+    position_size_units: float = 0.0
     estimated_risk_usd: float = 0.0
